rendir comisiones, falta la ultima parte

+++ Modelo/DetalleCompra.cs
@@ -11,8 +11,8 @@ namespace PalcoNet.Modelo
         public DateTime FechaCompra { get; set; }
         public String MetodoDePago { get; set; }
         public String Fila { get; set; }
-        public int Asiento { get; set; }
-        public int Precio { get; set; }
+        public Decimal Asiento { get; set; }
+        public Decimal Precio { get; set; }
         public String DescripcionUbicacion { get; set; }
         public String DescripcionPublicacion { get; set; }
         public DateTime FechaEvento { get; set; }
